調整controller 位置
